Move snapshotes to the __TEST__/Snapshot folder
