+++ src/__TEST__/Snapshots/Hall.test.js
@@ -1,16 +1,16 @@
 import React from 'react';
-import { render } from '@testing-library/react';
+import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
-import store from '../redux/store';
-import Halls from '../Pages/Halls/Halls';
+import store from '../../redux/store';
+import Hall from '../../Pages/Admin/Hall/index';
 
 describe('Login', () => {
   it('should render the login page', () => {
-    const app = render(
+    const app = renderer.create(
       <Provider store={store}>
         <Router>
-          <Halls />
+          <Hall />
         </Router>
       </Provider>,
     );
