+++ src/__TEST__/Snapshots/MyReservations.test.js
@@ -2,15 +2,15 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
-import store from '../redux/store';
-import Register from '../Pages/Auth/Register';
+import store from '../../redux/store';
+import MyReservations from '../../Pages/Users/MyReservations';
 
-describe('Register', () => {
-  it('should render the Register page', () => {
+describe('Login', () => {
+  it('should render the login page', () => {
     const app = render(
       <Provider store={store}>
         <Router>
-          <Register />
+          <MyReservations />
         </Router>
       </Provider>,
     );
