+++ src/__TEST__/Snapshots/Register.test.js
@@ -2,15 +2,15 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
-import store from '../redux/store';
-import Login from '../Pages/Auth/Login';
+import store from '../../redux/store';
+import Register from '../../Pages/Auth/Register';
 
-describe('Login', () => {
-  it('should render the login page', () => {
+describe('Register', () => {
+  it('should render the Register page', () => {
     const app = render(
       <Provider store={store}>
         <Router>
-          <Login />
+          <Register />
         </Router>
       </Provider>,
     );
